Modificacion en atributo puntosDeVida (conversion a double, porque espina resta el 5% y hay algunos que el 5% da numeros con coma). Ademas se debio modificar assertEquals(double,double) deprecated por assertEquals(double,double,double delta), donde delta es el rango aceptable, en nuestro caso 0D.

+++ src/fiuba/algo3/algoformers/modelo/Escenario/Superficies/EfectosSuperficie/EfectoSuperficieEspinas.java
@@ -10,7 +10,7 @@ public class EfectoSuperficieEspinas implements EfectoSuperficie {
     private int porcentaje = 5;
 
     public void aplicarEfecto(AlgoFormer algoformer){
-        int danio = algoformer.getPuntosDeVida() * this.porcentaje / 100;
+        double danio = algoformer.getPuntosDeVida() * this.porcentaje / 100;
         algoformer.recibirDanio(danio);
     }
